+++ R/write_xlsx.R
@@ -68,7 +68,7 @@ write_xlsx <- function(x, path = tempfile(fileext = ".xlsx"), col_names = TRUE,
                 header_offset = header_offset)
   sheets <- Map(function(el, df) .resolve_sheet_plan(el, df, reg, header_offset, props),
                 elems, dfs)
-  cm <- .resolve_constant_memory(dfs, props)
+  cm <- .resolve_constant_memory(dfs, props, sheets)
   ret <- .Call(C_write_data_frame_list, dfs, path, col_names, format_headers,
                use_zip64, reg$table, sheets, header_id,
                .properties_payload(props, cm$on))